Moved isNamedColor out of icon.tsx so it is unaffected by the icon testenv mock
